rename folder, separate components, assign Submit form to Save button inside Footer component
