+++ frontend/src/Pages/CreateForm/QuestionCard.module.css
@@ -1,0 +1,32 @@
+.questionCard {
+    display: flex;
+    align-items: center;
+    /* Alinha os itens verticalmente */
+    width: 80%;
+    border: 1px solid black;
+    border-radius: 20px;
+    padding: 10px;
+    gap: 7%;
+    overflow: hidden;
+    padding: 0 20px
+}
+
+#indexNumber {
+    max-width: fit-content;
+}
+
+.questionCard p {
+    flex: 1;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    /* Limits number of lines */
+    -webkit-box-orient: vertical;
+    /* Ensures vertical layout */
+}
+
+
+.questionCard button {
+    cursor: pointer;
+}
