+++ frontend/src/Pages/CreateForm/questionTypes.js
@@ -1,0 +1,17 @@
+const questionTypes = [{
+    val1ue: "TextQuestion",
+    label: 'Text'
+}, {
+    value: "MultipleChoiceQuestion",
+    label: 'Multiple Choice'
+},
+{
+    value: "SingleChoiceQuestion",
+    label: 'Single Choice'
+},
+{
+    value: "YesNoQuestion",
+    label: "Yes/No"
+}];
+
+export default questionTypes;
