--- frontend/src/Pages/CreateFormJose/Footer.module.css
@@ -1,30 +1,0 @@
-.saveButton {
-  width: 150px;
-  height: 40px;
-  margin: 8px;
-  border-radius: 14px;
-  text-align: center;
-  font-family: 'Space Mono', monospace;
-  font-size: 15px;
-  background-color: #a5a6a7;
-  color: white;
-  border: none;
-  cursor: pointer;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: fixed;
-  margin-bottom: 10px;
-  margin-right: 10px;
-}
-
-.footer {
-  height: 55px;
-  width: 100%;
-  background-color: #f0f0f0;
-  padding: 10px;
-  grid-area: footer;
-}
-
-/*<div style={{ display: 'flex', flexDirection: 'column', minHeight: '100vh' }}>
-<div style={{ marginBottom: '50px' }}>*/
