--- frontend/src/Pages/CreateFormJose/QuestionCard.module.css
@@ -1,16 +1,0 @@
-.questionCard {
-    display: flex;
-    justify-content: space-between;
-    width: auto;
-    border: 1px solid black;
-    border-radius: 20px;
-    padding: 0 20px;
-    /* 0 no topo e embaixo, 20px à direita e à esquerda */
-}
-
-.questionCard:hover {
-    display: flex;
-    background-color: blanchedalmond;
-    transition: 0.4s;
-    justify-content: space-between;
-}
