--- frontend/src/Pages/CreateFormJose/QuestionForm.module.css
@@ -1,5 +1,0 @@
-.questionChoice {
-    display: flex;
-    border: 1px solid black;
-    align-items: baseline;
-}
